core: create conflicts package
