+++ core/kt-osrd-conflicts/src/main/kotlin/fr/sncf/osrd/conflicts/IncrementalConflictDetector.kt
