+++ core/kt-osrd-conflicts/src/main/kotlin/fr/sncf/osrd/conflicts/Resources.kt
