+++ core/kt-osrd-conflicts/src/test/kotlin/fr/sncf/osrd/conflicts/IncrementalConflictDetectorTests.kt
