+++ core/settings.gradle
@@ -2,6 +2,7 @@ rootProject.name = 'osrd'
 include 'kt-fast-collections-annotations'
 include 'kt-fast-collections-generator'
 include 'kt-fast-collections'
+include 'kt-osrd-conflicts'
 include 'kt-osrd-utils'
 include 'kt-osrd-sncf-signaling'
 include 'kt-osrd-signaling'
